Update register.php
Changed `include()` to `require_once` because you only need to load them in once. Also removed the () because `include` and `require` are statements, not functions. `require` throws an exception which kills the page which is smarter when you're loading important files, where include only shows you a warning. You don't want to keep loading the page if the database, user and settings can't be loaded.